wayland: Make TabletToolV2Interface::sendPressure() take 0..1 pressure
65535 is an implementation detail of the tablet-v2 protocol.

+++ src/wayland/tablet_v2.h
@@ -128,7 +128,7 @@ public:
     void sendProximityOut();
     void sendUp();
     void sendDown();
-    void sendPressure(quint32 pressure);
+    void sendPressure(qreal pressure);
     void sendDistance(quint32 distance);
     void sendTilt(qreal degreesX, qreal degreesY);
     void sendRotation(qreal degrees);
